introduces new classes for maintability
